Address Xudong review comment 4: move readiness types to their own file
`RewriteReadiness`, `CandidateMetadata`, `ReadinessAnnotatedExec`, and
`readiness_from_plan` are now defined in a dedicated
`src/rewrite/readiness.rs` module, with the previous public paths
retained via re-exports so downstream callers don't have to change
their imports:

* `crate::materialized::RewriteReadiness` still resolves (re-exported).
* `crate::rewrite::exploitation::{CandidateMetadata,
  ReadinessAnnotatedExec, readiness_from_plan}` still resolve
  (re-exported).

The canonical home for every readiness-related type is now
`crate::rewrite::readiness`, so the surface a reader has to hold in
their head to understand the readiness contract lives in one place.

Test churn:
* `RewriteReadiness` sanity tests moved from materialized.rs to the
  new file's tests module.
* Added `readiness_annotated_exec_delegates_properties_and_children`
  in the new file to pin the wrapper's transparency (properties are
  the same Arc as the inner plan; the inner plan is the sole child).

Doc cross-references between modules use fully-qualified paths so the
rustdoc links still resolve after the move.

61/61 lib tests pass, fmt/clippy/doc clean.
