+++ src/rewrite.rs
@@ -21,6 +21,8 @@ pub mod exploitation;
 
 pub mod normal_form;
 
+pub mod readiness;
+
 mod util;
 
 extensions_options! {
