+++ src/rewrite/exploitation.rs
@@ -64,75 +64,18 @@ use datafusion_physical_expr::OrderingRequirements;
 use itertools::Itertools;
 use ordered_float::OrderedFloat;
 
-use crate::materialized::{cast_to_materialized, RewriteReadiness};
+use crate::materialized::cast_to_materialized;
 
 use super::normal_form::SpjNormalForm;
 use super::QueryRewriteOptions;
 
-/// Per-branch metadata inside a `OneOf` / `OneOfExec`. One variant per branch,
-/// aligned by index with the containing `branches` / `candidates` vector.
-///
-/// Lifecycle judgement (populated / unpopulated / stale / ...) is made in two
-/// stages via [`Materialized::rewrite_readiness`](crate::materialized::Materialized::rewrite_readiness):
-///
-/// 1. At LP-rewrite time, [`ViewMatcher`] drops `NotReady` providers and
-///    admits `Ready` + `Unknown` as candidates.
-/// 2. At physical-planning time, once DataFusion has invoked each provider's
-///    `scan()` (giving lazy indexes a chance to warm),
-///    [`ViewExploitationPlanner`] re-consults `rewrite_readiness()` and
-///    updates the metadata to the current value; providers that transitioned
-///    to `NotReady` between the two stages are dropped here.
-///
-/// So any `Materialized` variant that reaches the cost function reflects the
-/// **post-scan** readiness, and `NotReady` is guaranteed to have been filtered
-/// (upstream or downstream of scan) before the cost function runs. `Ready`
-/// and `Unknown` remain distinguishable so cost policy for the two can diverge
-/// (typically: trust `Ready` as safe to route, and fall back to a conservative
-/// estimate for `Unknown`).
-///
-/// The invariant enforced at construction inside [`ViewMatcher`]:
-///
-/// * Index 0 is always [`CandidateMetadata::Base`] — the query's original LP,
-///   with no MV rewrite applied. `OneOf::schema()` reads `branches[0].schema()`
-///   and therefore always exposes the query's schema (not an MV's).
-/// * Indices 1..N are [`CandidateMetadata::Materialized`] entries, sorted
-///   deterministically by `table_ref`. Sorting on the MV's registered
-///   `TableReference` (rather than on the branch LP) keeps the order stable
-///   under downstream LP transformations that happen between LP rewrite and
-///   physical planning — identity projection elimination, always-true filter
-///   folding, etc. change the LP but never the MV's registered name, so the
-///   alignment between this metadata and the physical candidate the cost
-///   function receives survives every optimizer pass.
-#[derive(Debug, Clone, PartialEq, PartialOrd, Eq, Hash)]
-pub enum CandidateMetadata {
-    /// The query's original branch (index 0 in every OneOf).
-    Base,
-    /// A materialized-view rewrite of the query. Only MVs that reported
-    /// `Ready` or `Unknown` readiness reach this state; `NotReady` MVs
-    /// are filtered upstream by [`ViewMatcher`]. Cost functions consult
-    /// the `readiness` field to distinguish the two cases: `Unknown`
-    /// (the trait default returned by [`RewriteReadiness`]) must not be
-    /// silently treated as `Ready`, since that would regress providers
-    /// that haven't declared a lifecycle.
-    Materialized {
-        /// Registered `TableReference` of the source MV. Stored directly
-        /// rather than as `.to_string()` so identifiers with dots or
-        /// quoting round-trip losslessly through the physical-time
-        /// catalog resolution. Stable across LP transformations; safe to
-        /// use as a sort key.
-        table_ref: TableReference,
-        /// Provider readiness as observed at physical-planning time — i.e.
-        /// after DataFusion has invoked `TableProvider::scan()` on this
-        /// branch. Set once at LP rewrite from the provider's initial
-        /// report and refreshed from the same provider inside
-        /// [`ViewExploitationPlanner::plan_extension`] so lazy-index
-        /// providers surface their warmed-up value. Downstream cost
-        /// functions branch on this to implement their `Unknown` policy
-        /// (e.g. fall back to the base scan cost rather than trust an
-        /// EmptyExec as predicate-pruned).
-        readiness: RewriteReadiness,
-    },
-}
+// Bring readiness types in for use inside this module AND re-export
+// them for backward compatibility. Callers that used to import from
+// `crate::rewrite::exploitation` (the pre-move location) keep working;
+// the canonical home is now `crate::rewrite::readiness`.
+pub use super::readiness::{
+    readiness_from_plan, CandidateMetadata, ReadinessAnnotatedExec, RewriteReadiness,
+};
 
 /// Logical rewrite metadata propagated alongside equivalent candidate plans.
 #[derive(Debug, Clone, Default, PartialEq, PartialOrd, Eq, Hash)]
@@ -1044,129 +987,6 @@ impl PhysicalOptimizerRule for PruneCandidates {
     }
 }
 
-/// Wraps the `ExecutionPlan` returned by a `Materialized` provider's
-/// `scan()` together with the readiness value captured at scan time.
-///
-/// Providers use this to close the race between "which snapshot did
-/// scan read from" and "what does `rewrite_readiness()` return now".
-/// Sampling `rewrite_readiness()` again after `scan()` has returned is
-/// racy — a concurrent snapshot swap can publish new state between the
-/// two calls, letting an `EmptyExec` from the old snapshot be labelled
-/// with readiness from the new one. If the provider computes both
-/// atomically (from the same state Arc, under the same read lock, etc.)
-/// and wraps the returned plan with `ReadinessAnnotatedExec`, the
-/// physical-time refresh in `ViewExploitationPlanner::plan_extension`
-/// reads the annotated value verbatim instead of re-sampling.
-///
-/// The wrapper is transparent: all `ExecutionPlan` methods delegate to
-/// the inner plan, so downstream operators see the same shape as if the
-/// provider returned `inner` directly.
-///
-/// Providers that don't wrap fall back to the pre-existing (racy)
-/// sampling path in the refresh, which remains supported for backward
-/// compatibility.
-#[derive(Debug)]
-pub struct ReadinessAnnotatedExec {
-    inner: Arc<dyn ExecutionPlan>,
-    readiness: RewriteReadiness,
-    properties: Arc<PlanProperties>,
-}
-
-impl ReadinessAnnotatedExec {
-    /// Wrap `inner` with the readiness captured atomically at scan time.
-    pub fn new(inner: Arc<dyn ExecutionPlan>, readiness: RewriteReadiness) -> Self {
-        let properties = Arc::clone(inner.properties());
-        Self {
-            inner,
-            readiness,
-            properties,
-        }
-    }
-
-    /// Readiness captured at the moment the provider produced `inner`.
-    pub fn readiness(&self) -> RewriteReadiness {
-        self.readiness
-    }
-
-    /// The wrapped plan.
-    pub fn inner(&self) -> &Arc<dyn ExecutionPlan> {
-        &self.inner
-    }
-}
-
-impl DisplayAs for ReadinessAnnotatedExec {
-    fn fmt_as(&self, t: DisplayFormatType, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match t {
-            DisplayFormatType::Default | DisplayFormatType::Verbose => {
-                write!(f, "ReadinessAnnotatedExec: readiness={:?}", self.readiness)
-            }
-            DisplayFormatType::TreeRender => Ok(()),
-        }
-    }
-}
-
-impl ExecutionPlan for ReadinessAnnotatedExec {
-    fn name(&self) -> &str {
-        "ReadinessAnnotatedExec"
-    }
-
-    fn properties(&self) -> &Arc<PlanProperties> {
-        &self.properties
-    }
-
-    fn children(&self) -> Vec<&Arc<dyn ExecutionPlan>> {
-        vec![&self.inner]
-    }
-
-    fn with_new_children(
-        self: Arc<Self>,
-        children: Vec<Arc<dyn ExecutionPlan>>,
-    ) -> Result<Arc<dyn ExecutionPlan>> {
-        if children.len() != 1 {
-            return Err(DataFusionError::Plan(format!(
-                "ReadinessAnnotatedExec expects exactly one child, got {}",
-                children.len()
-            )));
-        }
-        Ok(Arc::new(ReadinessAnnotatedExec::new(
-            children.into_iter().next().unwrap(),
-            self.readiness,
-        )))
-    }
-
-    fn execute(
-        &self,
-        partition: usize,
-        context: Arc<TaskContext>,
-    ) -> Result<SendableRecordBatchStream> {
-        self.inner.execute(partition, context)
-    }
-
-    fn partition_statistics(
-        &self,
-        partition: Option<usize>,
-    ) -> Result<Arc<datafusion_common::Statistics>> {
-        self.inner.partition_statistics(partition)
-    }
-}
-
-/// Depth-first search for the nearest `ReadinessAnnotatedExec` in a
-/// physical plan tree. Returns its readiness value, or `None` if the
-/// tree doesn't contain one (e.g. the provider hasn't opted in). Used
-/// by the physical-time refresh to prefer atomically-captured readiness
-/// over racy re-sampling.
-pub fn readiness_from_plan(plan: &Arc<dyn ExecutionPlan>) -> Option<RewriteReadiness> {
-    if let Some(annotated) = plan.downcast_ref::<ReadinessAnnotatedExec>() {
-        return Some(annotated.readiness());
-    }
-    for child in plan.children() {
-        if let Some(readiness) = readiness_from_plan(child) {
-            return Some(readiness);
-        }
-    }
-    None
-}
-
 /// Compare two Arrow schemas ignoring field nullability.
 fn schemas_equal_ignoring_nullability(a: &arrow_schema::Schema, b: &arrow_schema::Schema) -> bool {
     a.fields().len() == b.fields().len()
@@ -2076,7 +1896,8 @@ mod tests_view_matcher_readiness_filter {
 mod tests_physical_time_readiness_refresh {
     use super::tests_view_matcher_readiness_filter::*;
     use super::*;
-    use crate::materialized::{register_materialized, RewriteReadiness};
+    use crate::materialized::register_materialized;
+    use crate::rewrite::readiness::ReadinessAnnotatedExec;
     use datafusion::execution::context::SessionContext;
     use datafusion_expr::LogicalPlanBuilder;
     use std::sync::Arc;
